PR-AN-06: Refactor unit tests for apis into separate files.
