+++ lib/api/choices.js
@@ -3,7 +3,7 @@ const argsert = require('../argsert')
 module.exports = function (options, context, populateParserHintObject) {
   return function (key, value) {
     argsert('<object|string|array> [string|array]', [key, value], arguments.length)
-    populateParserHintObject(this.choices, true, 'choices', key, value)
-    return this
+    populateParserHintObject(context.choices, true, 'choices', key, value)
+    return context
   }
 }
