+++ package.json
@@ -43,7 +43,7 @@
   },
   "scripts": {
     "posttest": "standard",
-    "test": "nyc --cache mocha --require ./test/before.js --timeout=8000 --check-leaks",
+    "test": "nyc --cache mocha --require ./test/before.spec.js './test/**/*.spec.js' --timeout=8000 --check-leaks",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "release": "standard-version"
   },
