+++ test/command.spec.js
