+++ test/completion.spec.js
