+++ test/usage.spec.js
