+++ test/yargs.spec.js
@@ -537,71 +537,6 @@ describe('yargs dsl tests', () => {
     })
   })
 
-  describe('number', () => {
-    it('accepts number arguments when a number type is specified', () => {
-      const argv = yargs('-w banana')
-        .number('w')
-        .argv
-
-      expect(typeof argv.w).to.equal('number')
-    })
-
-    it('should expose an options short-hand for numbers', () => {
-      const argv = yargs('-w banana')
-        .option('w', {
-          number: true
-        })
-        .alias('w', 'x')
-        .argv
-
-      expect(typeof argv.w).to.equal('number')
-      expect(typeof argv.x).to.equal('number')
-    })
-  })
-
-  describe('choices', () => {
-    it('accepts an object', () => {
-      const optChoices = yargs([])
-        .choices({
-          color: ['red', 'green', 'blue'],
-          stars: [1, 2, 3, 4, 5]
-        })
-        .choices({
-          size: ['xl', 'l', 'm', 's', 'xs']
-        })
-        .getOptions().choices
-
-      optChoices.should.deep.equal({
-        color: ['red', 'green', 'blue'],
-        stars: [1, 2, 3, 4, 5],
-        size: ['xl', 'l', 'm', 's', 'xs']
-      })
-    })
-
-    it('accepts a string and array', () => {
-      const optChoices = yargs([])
-        .choices('meat', ['beef', 'chicken', 'pork', 'bison'])
-        .choices('temp', ['rare', 'med-rare', 'med', 'med-well', 'well'])
-        .getOptions().choices
-
-      optChoices.should.deep.equal({
-        meat: ['beef', 'chicken', 'pork', 'bison'],
-        temp: ['rare', 'med-rare', 'med', 'med-well', 'well']
-      })
-    })
-
-    it('accepts a string and single value', () => {
-      const optChoices = yargs([])
-        .choices('gender', 'male')
-        .choices('gender', 'female')
-        .getOptions().choices
-
-      optChoices.should.deep.equal({
-        gender: ['male', 'female']
-      })
-    })
-  })
-
   describe('locale', () => {
     it('uses english as a default locale', () => {
       ['LANGUAGE', 'LC_ALL', 'LANG', 'LC_MESSAGES'].forEach((e) => {
@@ -1333,80 +1268,6 @@ describe('yargs dsl tests', () => {
     })
   })
 
-  describe('normalize', () => {
-    it('normalizes paths passed as arguments', () => {
-      const argv = yargs('--path /foo/bar//baz/asdf/quux/..')
-        .normalize(['path'])
-        .argv
-
-      argv.path.should.equal(['', 'foo', 'bar', 'baz', 'asdf'].join(path.sep))
-    })
-
-    it('normalizes path when when it is updated', () => {
-      const argv = yargs('--path /batman')
-        .normalize(['path'])
-        .argv
-
-      argv.path = '/foo/bar//baz/asdf/quux/..'
-      argv.path.should.equal(['', 'foo', 'bar', 'baz', 'asdf'].join(path.sep))
-    })
-
-    it('allows key to be specified with option shorthand', () => {
-      const argv = yargs('--path /batman')
-        .option('path', {
-          normalize: true
-        })
-        .argv
-
-      argv.path = '/foo/bar//baz/asdf/quux/..'
-      argv.path.should.equal(['', 'foo', 'bar', 'baz', 'asdf'].join(path.sep))
-    })
-
-    it('can be disabled with option shorthand', () => {
-      const argv = yargs('--path /batman')
-        .option('path', {
-          normalize: false
-        })
-        .argv
-
-      argv.path = 'mongodb://url'
-      argv.path.should.equal('mongodb://url')
-    })
-  })
-
-  describe('narg', () => {
-    it('accepts a key as the first argument and a count as the second', () => {
-      const argv = yargs('--foo a b c')
-        .nargs('foo', 2)
-        .argv
-
-      argv.foo.should.deep.equal(['a', 'b'])
-      argv._.should.deep.equal(['c'])
-    })
-
-    it('accepts a hash of keys and counts', () => {
-      const argv = yargs('--foo a b c')
-        .nargs({
-          foo: 2
-        })
-        .argv
-
-      argv.foo.should.deep.equal(['a', 'b'])
-      argv._.should.deep.equal(['c'])
-    })
-
-    it('allows key to be specified with option shorthand', () => {
-      const argv = yargs('--foo a b c')
-        .option('foo', {
-          nargs: 2
-        })
-        .argv
-
-      argv.foo.should.deep.equal(['a', 'b'])
-      argv._.should.deep.equal(['c'])
-    })
-  })
-
   describe('global', () => {
     it('does not reset a global options when reset is called', () => {
       const y = yargs('--foo a b c')
@@ -1613,55 +1474,6 @@ describe('yargs dsl tests', () => {
     })
   })
 
-  describe('skipValidation', () => {
-    it('skips validation if an option with skipValidation is present', () => {
-      const argv = yargs(['--koala', '--skip'])
-          .demand(1)
-          .fail((msg) => {
-            expect.fail()
-          })
-          .skipValidation(['skip', 'reallySkip'])
-          .argv
-      argv.koala.should.equal(true)
-    })
-
-    it('does not skip validation if no option with skipValidation is present', (done) => {
-      const argv = yargs(['--koala'])
-          .demand(1)
-          .fail(msg => done())
-          .skipValidation(['skip', 'reallySkip'])
-          .argv
-      argv.koala.should.equal(true)
-    })
-
-    it('allows key to be specified with option shorthand', () => {
-      const argv = yargs(['--koala', '--skip'])
-          .demand(1)
-          .fail((msg) => {
-            expect.fail()
-          })
-          .option('skip', {
-            skipValidation: true
-          })
-          .argv
-      argv.koala.should.equal(true)
-    })
-
-    it('allows having an option that skips validation but not skipping validation if that option is not used', () => {
-      let skippedValidation = true
-      yargs(['--no-skip'])
-          .demand(5)
-          .option('skip', {
-            skipValidation: true
-          })
-          .fail((msg) => {
-            skippedValidation = false
-          })
-          .argv
-      expect(skippedValidation).to.equal(false)
-    })
-  })
-
   describe('.help()', () => {
     it('enables `--help` option and `help` command without arguments', () => {
       const option = checkOutput(() => yargs('--help')
@@ -1786,150 +1598,6 @@ describe('yargs dsl tests', () => {
         '  -h, --help, --info  Show help  [boolean]'
       ])
       h.result.should.have.property('_').and.deep.equal(['h'])
-    })
-  })
-
-  describe('.coerce()', () => {
-    it('supports string and function args (as option key and coerce function)', () => {
-      const argv = yargs(['--file', path.join(__dirname, 'fixtures', 'package.json')])
-        .coerce('file', arg => JSON.parse(fs.readFileSync(arg, 'utf8')))
-        .argv
-      expect(argv.file).to.have.property('version').and.equal('9.9.9')
-    })
-
-    it('supports object arg (as map of multiple options)', () => {
-      const argv = yargs('--expand abc --range 1..3')
-        .coerce({
-          expand (arg) {
-            return arg.split('')
-          },
-          range (arg) {
-            const arr = arg.split('..').map(Number)
-            return { begin: arr[0], end: arr[1] }
-          }
-        })
-        .argv
-      expect(argv.expand).to.deep.equal(['a', 'b', 'c'])
-      expect(argv.range).to.have.property('begin').and.equal(1)
-      expect(argv.range).to.have.property('end').and.equal(3)
-    })
-
-    it('supports array and function args (as option keys and coerce function)', () => {
-      const argv = yargs(['--src', 'in', '--dest', 'out'])
-        .coerce(['src', 'dest'], arg => path.resolve(arg))
-        .argv
-      argv.src.should.match(/in/).and.have.length.above(2)
-      argv.dest.should.match(/out/).and.have.length.above(3)
-    })
-
-    it('allows an error to be handled by fail() handler', () => {
-      let msg
-      let err
-      let jsonErrMessage
-      yargs('--json invalid')
-        .coerce('json', (arg) => {
-          try {
-            JSON.parse(arg)
-          } catch (err) {
-            jsonErrMessage = err.message
-          }
-          return JSON.parse(arg)
-        })
-        .fail((m, e) => {
-          msg = m
-          err = e
-        })
-        .argv
-      expect(msg).to.equal(jsonErrMessage)
-      expect(err).to.exist
-    })
-
-    it('supports an option alias', () => {
-      const argv = yargs('-d 2016-08-12')
-        .coerce('date', Date.parse)
-        .alias('date', 'd')
-        .argv
-      argv.date.should.equal(1470960000000)
-    })
-
-    it('supports a global option within command', () => {
-      let regex
-      yargs('check --regex x')
-        .global('regex')
-        .coerce('regex', RegExp)
-        .command('check', 'Check something', {}, (argv) => {
-          regex = argv.regex
-        })
-        .argv
-      expect(regex).to.be.an.instanceof(RegExp)
-      regex.toString().should.equal('/x/')
-    })
-
-    it('is supported by .option()', () => {
-      const argv = yargs('--env SHELL=/bin/bash')
-        .option('env', {
-          coerce (arg) {
-            const arr = arg.split('=')
-            return { name: arr[0], value: arr[1] || '' }
-          }
-        })
-        .argv
-      expect(argv.env).to.have.property('name').and.equal('SHELL')
-      expect(argv.env).to.have.property('value').and.equal('/bin/bash')
-    })
-
-    it('supports positional and variadic args for a command', () => {
-      let age
-      let dates
-      yargs('add 30days 2016-06-13 2016-07-18')
-        .command('add <age> [dates..]', 'Testing', yargs => yargs
-            .coerce('age', arg => parseInt(arg, 10) * 86400000)
-            .coerce('dates', arg => arg.map(str => new Date(str))), (argv) => {
-              age = argv.age
-              dates = argv.dates
-            })
-        .argv
-      expect(age).to.equal(2592000000)
-      expect(dates).to.have.lengthOf(2)
-      dates[0].toString().should.equal(new Date('2016-06-13').toString())
-      dates[1].toString().should.equal(new Date('2016-07-18').toString())
-    })
-
-    it('returns camelcase args for a command', () => {
-      let age1
-      let age2
-      let dates
-      yargs('add 30days 2016-06-13 2016-07-18')
-        .command('add <age-in-days> [dates..]', 'Testing', yargs => yargs
-            .coerce('age-in-days', arg => parseInt(arg, 10) * 86400000)
-            .coerce('dates', arg => arg.map(str => new Date(str))), (argv) => {
-              age1 = argv.ageInDays
-              age2 = argv['age-in-days']
-              dates = argv.dates
-            })
-        .argv
-      expect(age1).to.equal(2592000000)
-      expect(age2).to.equal(2592000000)
-      expect(dates).to.have.lengthOf(2)
-      dates[0].toString().should.equal(new Date('2016-06-13').toString())
-      dates[1].toString().should.equal(new Date('2016-07-18').toString())
-    })
-
-    it('allows an error from positional arg to be handled by fail() handler', () => {
-      let msg
-      let err
-      yargs('throw ball')
-        .command('throw <msg>', false, yargs => yargs
-            .coerce('msg', (arg) => {
-              throw new Error(arg)
-            })
-            .fail((m, e) => {
-              msg = m
-              err = e
-            }))
-        .argv
-      expect(msg).to.equal('ball')
-      expect(err).to.exist
     })
   })
